Testing barrel deletion
