+++ src/browser/bootstrap.spec.ts
@@ -1,0 +1,4 @@
+import {BrowserDynamicTestingModule, platformBrowserDynamicTesting} from '@angular/platform-browser-dynamic/testing';
+import { TestBed } from '@angular/core/testing';
+
+TestBed.initTestEnvironment(BrowserDynamicTestingModule, platformBrowserDynamicTesting());
