+++ src/browser/stores/index.ts
@@ -2,4 +2,4 @@
  * @module browser
  */
 /** End Typedoc Module Declaration */
-export * from './http.store';
+// export * from './http.store';
