+++ src/common/exceptions/index.ts
@@ -2,4 +2,4 @@
  * @module common
  */
 /** End Typedoc Module Declaration */
-export * from './exceptions';
+// export * from './exceptions';
