+++ src/common/models/index.ts
@@ -2,7 +2,7 @@
  * @module common
  */
 /** End Typedoc Module Declaration */
-export * from './model';
-export * from './collection';
-export * from './types';
-export * from './relations';
+// export * from './model';
+// export * from './collection';
+// export * from './types';
+// export * from './relations';
