+++ src/common/models/types/index.ts
@@ -2,6 +2,6 @@
  * @module common
  */
 /** End Typedoc Module Declaration */
-export * from './primary.decorator';
-export * from './storedProperty.decorator';
-export * from './timestamp.decorator';
+// export * from './primary.decorator';
+// export * from './storedProperty.decorator';
+// export * from './timestamp.decorator';
