+++ src/common/registry/index.ts
@@ -2,5 +2,5 @@
  * @module common
  */
 /** End Typedoc Module Declaration */
-export * from './entityRegistry';
-export * from './decorators';
+// export * from './entityRegistry';
+// export * from './decorators';
