+++ src/server/bootstrap/controllers.bootstrapper.spec.ts
@@ -2,11 +2,11 @@ import { TestBed } from '@angular/core/testing';
 import { Injectable } from '@angular/core';
 import { RemoteCliMock } from '../services/remoteCli.service.mock';
 import { RemoteCli } from '../services/remoteCli.service';
-import { ServerMock } from '../servers/abstract.server.spec';
+import { ServerMock } from '../servers/abstract.server.mock';
 import { Logger } from '../../common/services/logger.service';
 import { LoggerMock } from '../../common/services/logger.service.mock';
 import { Server } from '../servers/abstract.server';
-import { bootstrap, BootstrapResponse } from './index';
+import { bootstrap, BootstrapResponse } from './bootstrap';
 import { registry } from '../../common/registry/entityRegistry';
 import { AbstractController } from '../controllers/abstract.controller';
 import { Route } from '../controllers/route.decorator';
