+++ src/server/bootstrap/index.ts
@@ -4,4 +4,4 @@
 /**
  * Barrel module only for exporting public implementations of bootstrap interfaces and functions
  */
-export * from './bootstrap';
+// export * from './bootstrap';
