+++ src/server/bootstrap/services.bootstrapper.spec.ts
@@ -2,12 +2,12 @@ import { TestBed } from '@angular/core/testing';
 import { Injectable } from '@angular/core';
 import { Logger } from '../../common/services/logger.service';
 import { LoggerMock } from '../../common/services/logger.service.mock';
-import { bootstrap, BootstrapResponse } from './index';
+import { bootstrap, BootstrapResponse } from './bootstrap';
 import { registry } from '../../common/registry/entityRegistry';
 import { AbstractService } from '../../common/services/service';
 import { RemoteCliMock } from '../services/remoteCli.service.mock';
 import { RemoteCli } from '../services/remoteCli.service';
-import { ServerMock } from '../servers/abstract.server.spec';
+import { ServerMock } from '../servers/abstract.server.mock';
 import { Server } from '../servers/abstract.server';
 import Spy = jasmine.Spy;
 import { AuthServiceMock } from '../services/auth.service.mock';
