+++ src/server/index.ts
@@ -14,12 +14,12 @@
  * @preferred
  */
 /** */
-export * from './main';
-export * from './bootstrap';
-export * from './controllers';
-export * from './servers';
-export * from './services';
-export * from './seeders';
-export * from './migrations';
-export * from './stores';
-export * from './middleware';
+// export * from './main';
+// export * from './bootstrap';
+// export * from './controllers';
+// export * from './servers';
+// export * from './services';
+// export * from './seeders';
+// export * from './migrations';
+// export * from './stores';
+// export * from './middleware';
