+++ src/server/servers/abstract.server.spec.ts
@@ -1,7 +1,6 @@
 import { Server, RouteConfig } from './abstract.server';
 import { RemoteCli } from '../services/remoteCli.service';
 import { Logger } from '../../common/services/logger.service';
-import { Injectable } from '@angular/core';
 import { inject, TestBed, async } from '@angular/core/testing';
 import { LoggerMock } from '../../common/services/logger.service.mock';
 import { RemoteCliMock } from '../services/remoteCli.service.mock';
@@ -9,35 +8,7 @@ import { Response } from '../controllers/response';
 import Spy = jasmine.Spy;
 import { AuthService } from '../services/auth.service';
 import { AuthServiceMock } from '../services/auth.service.mock';
-
-@Injectable()
-export class ServerMock extends Server {
-
-  public getEngine(): any {
-    return undefined;
-  }
-
-  constructor(logger: Logger, remoteCli: RemoteCli) {
-    super(logger, remoteCli);
-  }
-
-  protected registerRouteWithEngine(config: RouteConfig): this {
-    return this;
-  }
-
-  protected initialize(): this {
-    return this;
-  }
-
-  public startEngine(): Promise<this> {
-    return Promise.resolve(this);
-  }
-
-  public registerStaticLoader(webroot?: string): this {
-    return this;
-  }
-
-}
+import { ServerMock } from './abstract.server.mock';
 
 describe('Server', () => {
